api/requirements.txt
alembic
email_validator
fastapi
pwdlib[argon2]
pydantic
python-decouple
python-dotenv
python-multipart
psycopg2-binary
httpx
SQLAlchemy
sqlmodel
uvicorn[standard]